Update composer.json
The new version of the spam protection module has a requirement for the framework already, so this causes issues trying to install this module via composer as there are framework conflictions and it won't install.

Anyway we can remove this line so it'll install? And now thinking if this works with the new spam protection API changes.